requirements.txt
anthropic>=0.25.0
redis>=5.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
certifi>=2024.0.0
